Unify reasoning and stochastic services into stochastic-clarity
