--- services/clear-thought/go.mod
@@ -1,22 +1,0 @@
-module github.com/c3mb0/cemcp/services/clear-thought
-
-go 1.24.5
-
-require (
-	github.com/c3mb0/cemcp/pkg v0.0.0
-	github.com/mark3labs/mcp-go v0.37.0
-)
-
-require (
-	github.com/bahlo/generic-list-go v0.2.0 // indirect
-	github.com/buger/jsonparser v1.1.1 // indirect
-	github.com/google/uuid v1.6.0 // indirect
-	github.com/invopop/jsonschema v0.13.0 // indirect
-	github.com/mailru/easyjson v0.7.7 // indirect
-	github.com/spf13/cast v1.7.1 // indirect
-	github.com/wk8/go-ordered-map/v2 v2.1.8 // indirect
-	github.com/yosida95/uritemplate/v3 v3.0.2 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
-
-replace github.com/c3mb0/cemcp/pkg => ../../pkg
